+++ src/traits/i-active-items/interface.ts
@@ -21,6 +21,11 @@ export interface Item extends Dictionary {
 	 * True if the item is active
 	 */
 	active?: boolean;
+
+	/**
+	 * True if item can possibly be active
+	 */
+	activatable?: boolean;
 }
 
 export type Active = unknown | Set<unknown>;