+++ src/base/b-router/CHANGELOG.md
@@ -15,9 +15,27 @@ Changelog
 
 * Added new event `hrefTransition` to `b-router` to provide the capability to prevent router transition upon link click
 
+## v3.51.1 (2023-06-27)
+
+#### :bug: Bug Fix
+
+* Handle unsuitable `pathParams` values in the `fillRouteParams` function
+
+## v3.51.0 (2023-06-27)
+
+#### :rocket: New Feature
+
+* Added support for `mailto:` and `tel:` href-s
+
+## v3.47.1 (2023-05-18)
+
+#### :bug: Bug Fix
+
+* Replace `undefined` values in `route.params` by an alias or query param, if necessary
+
 ## v3.44.3 (2023-03-30)
 
-#### :bug: [Bug Fix]
+#### :bug: Bug Fix
 
 * Overriding original parameter by alias in route